SPM: boot OP-TEE as a guest Secure Partition on top of Hafnium
Add an SPM scenario where Hafnium is loaded as the SPMC at S-EL2
and OP-TEE is instanciated as a Secure Partition at S-EL1.
System boots Hafnium and OP-TEE in the SWd. OP-TEE is booted as
the first and unique Secure Partition. Later when TFTF is booted
in the NWd, the PSA FF-A direct messaging test perform a probing
of OP-TEE's version. The direct message is relayed from TFTF(NS-EL2)
to SPMD(EL3) to Hafnium(S-EL2) and down to OP-TEE(S-EL1). OP-TEE
replies with a direct message response on the way back to TFTF.

Change-Id: Iab033af3d933537df80cc16564bf4172c72f0f93
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>