Add more build configurations for Coverity MISRA checks
This patch adds more configs for Coverity MISRA checks.
It covers more source files that are not covered right now.

Unlike the MISRA checks we have now, these are nominated,
i.e. they are only run on patches that touch specific source files.
Therefore, this patch also updates the nomination rules file.
I chose to create a new "tf-l2-coverity-misra-nominated" group to
differentiate them from the ones that are always run.

An ultimate goal is to move the configs (at least some of them) in the
"tf-l2-coverity-misra" group to the new nominated group as we don't
need to run all of them on every patch.

Change-Id: I1939d075e587732c6e6239f509eeeb8e5afeac4e
Signed-off-by: Zelalem <zelalem.aweke@arm.com>