Avoid cloning code coverage when it's not used
This works by being explicit that we don't want to clone the code
coverage repo by setting the bash variable `import_cc` to 1. `import_cc`
is directly assigned to `no_cc`, and setting `no_cc` to 1 indicates
"I'm not running coverage".

This also informs the user (in green) that code coverage is disabled as
they have requested.

Change-Id: Iac331ec65f568cfe13b8a3fa9fa0146a630a8ae7