FVP Linux and TFTF YAML Support
Introduce support for fvp-[linux|tftf]-yaml generation, the latter to be
consumed by an FVP LAVA device on behalf of TF-A CI (Jenkins).

Two build configs and one run config are instrumented for RFC
purposes, but once patch is more mature, many more will come.

This work is part of M2 (FVP Support (TF-M ID3, TF-A ID26)), and
M3 (Run TF-TF tests "bare metal" on FVP (TF-A ID24)) of [1].

Note: Testing and review is on-going so patchset may required some
changes if needed.

[1] https://developer.trustedfirmware.org/w/collaboration/openci/

Change-Id: I613cd2e4028f6cef3e96976cf5f3f1180e156440
Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>