Fix fetch_file() function
Make sure that the 'sa' variable is defined after the 'saveas' variable,
which it depends on.

Otherwise, invokations of fetch_file() which do not specify any 'saveas'
variable get an empty 'sa' variable. This wasn't a big problem back when
we were using wget to fetch the file but it is now that we've switched
to curl. curl writes its output to stdout by default so it is crucial to
pass it the '-o' option (through the 'sa' variable) to get the desired
behaviour.

Change-Id: I2d8c0bfbf175f0fb64b61195c5e3cb09ac573f27
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>