Add support for environmental settings
Introduce a script that overrides the relevant environment variables to
work on the Arm CI infrastructure. It is automatically sourced from
utils.sh if we are running on Arm's Jenkins server.

If running the scripts on a local machine, the user is expected to
provide their own environmental settings file through the host_env
variable if needed.

Change-Id: Ibf219ab10988ed97731fb17434ff01b34fdf93aa
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>