utils.sh: Fix detection of Arm infrastructure
This patch fixes a bug introduced in patch ae11ebfe35c7e ("Add support
for environmental settings"). In utils.sh, we want to know whether the
JENKINS_URL environment variable contains "arm.com" but we are not
quiescing the output of the grep command. This messes up some scripts
that source utils.sh and redirect their output to some file, which now
contains the spurious "jenkins.oss.arm.com" string.

We are only interested in the exit code of the grep command so quiesce
its output.

Change-Id: Ia47f1b4d1f150db1aa346f2eb2ec95fa9b9a2346
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>