Introduce yaml-docker-config python script
Script that takes a FVP yaml file and produces the parameters for a docker-run
command. Conceptually, this is similar to 'pgk-config' system application but
this scripts applies to docker-run parameters.

To exemplify its usage, let's assume we have the a FVP yaml file at ~/fvp.yaml
then launch the container using the script to provide the correct model parameters

   $ docker run `./yaml-docker-config ~/fvp.yaml`

If no errors, the (containerized) FVP model should be up and running. In case the
docker image is not found, please create one following the instructions on the
fvp/README.md file.

Change-Id: I03aab0b882d4b0204df90b2b65cbdba637e19a51
Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>