don't provide juno-revision tag to LAVA
juno-revision tag is passed to LAVA which decides the board on
which tests will run, currently it is set R0. This causes a bottleneck
as we have different revisions of juno boards(R0/R1/R2) available in the
farm but jenkins job wait for R0 boards to be free and available.

This patch gets rid of providing a specific juno revision to LAVA as
TF-A CI does not need it.

Change-Id: Ibd949940ecd0dceaf04f02a218439044be90972a
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>