Add support to test TF-A Measured Boot through an external TPM service.
Perform automatic tests to ensure that Measured Boot functionality on
TF-A can interact with an external TPM service.

This patch performs the following tests:

1.- Check that PCR0 Digest is not all zeros
2.- Check that the rest of PCRs are all zero
3.- Check that the event log written by TF-A is the same as read by
    the TPM service.

The patch is not meant to test whether the digests are correct or not
(testing if the TPM service is buggy or not is beyond the scope of
this test). It just tests that an external TPM service is able to
access the TPM event log generated by TF-A and that it can be properly
processed.

Change-Id: I5fba87005886ff549345bd92675d2f2a9fe44e79
Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>