Do a fresh clone when syncing repos
We have been having issues while trying to rename
tags in the tf.org CI repo. We believe this is due to
the sync script using old copy of the repo.

To fix this do a fresh clone of the repos for now.
Might revert back this change once the issue is fixed.

Change-Id: I13eb516d0e3aa460e1e72350c8c9175443972765
Signed-off-by: Zelalem <zelalem.aweke@arm.com>