Get utility script name from TF configuration
The problem was found with CI build scripts when building rdn1edgex2
platform. Both rdn1edge and rdn1edgex2 have PLAT=rdn1edge in their TF
configs (which is correct, as there is no rdn1edgex2 build option
for TF-A).The problem is that PLAT sets 'plat' variable which is used
to call ${plat}_utils.sh and rdn1edge_utils.sh is called instead of
rdn1edgex2_utils.sh.

To resolve this issue introduced a new config parameter “SCRIPT_NAME”
inside the TF configuration which is used to source the utility script
for e.g. rdn1edge_utils.sh.

Change-Id: I5c50f71acc77f0bb2d24393e784360adb36adf95
Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>