fvp: Disable telnet ports for UARTs 1, 2 and 3
LAVA can only read one UART, currently, that is UART 0. Some tests
output to UARTs 1, 2 or 3 in addition to the main UART. If these UARTs
are not read, the model can hang once the buffers are full.

As a temporary measure until LAVA supports reading multiple UARTs,
disable the extra UARTs on tf.validation.linaro.org, but keep it enabled
for Arm internal infrastructure.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Change-Id: Ibea2fb0f80e54129c0e84657ed437cf540509de6