Bug fix in the expect script deployed by the fvp-aemv8a.tbb.disable_dyn_auth run fragment.
Prior to this patch, the expect script didn't watch the strings
emitted by the TFTF to the UART0 port. Hence, even when all the
tests in the TFTF test suite finished successfully, the expect
script didn't terminate the model in a clean way, rather a timeout
led to the model termination.

This patch adds a new expect script which broadly tracks two patterns
that are put to the UART0 console by the CPUs:
  1. Related to the disabling of image authorization dynamically
  2. Booting of BL32(TFTF image) and completion of tests

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I20a6bcd6a0134de064ceef6e8999f87a6c4306b3