Download a prebuilt version of OPTEE binary
As explained in commit 284bd00e3ed752a45f90add999261c2da6d098c5, we
cannot use the prebuilt OPTEE binary shipped with the latest Linaro
release (20.01).

Instead, we've deployed on the CI servers a prebuilt binary built from
a more recent version of OPTEE (version 3.7.0, commit de5333ed), which
has the required patch for the Base AEM RevC FVP. Note we could have
used an even more recent version of OPTEE (the latest release is 3.12)
but this one corresponds to what we've been using in the ArmCI for a
while so it was just simpler to keep things as is.

Change-Id: I6ed1843616ee4b179d57aec36ee25961c3c5eeb1
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>