script: fix hardcoded cmsis path
At present, clone_repos.sh for the SCP repository and to update
cmsis submodule in it, changes directory using 'cd cmsis'.
This is hardcoded and can break if the path is changed by the
SCP firmware repository.

This change uses 'git submodule status' command to get the exact
location of the submodule so that the script will continue to work
on the cmsis irrespective of the location availabe in the SCP firmware
repo.

Signed-off-by: Girish Pathak <girish.pathak@arm.com>
Change-Id: Ie3b9e9027c3d38196af261e15e02bb7f7bb716c5