docker/fvp: use multi-stage approach and other minor changes
Changes include: 1. create a multi-stage image, first image for
installing FVP dependencies and second image from model installation
2. Avoid mkdir on MODEL_DIR 3. Include comment about fvp models: some models
contain an installer and some do not 4. Run the installer together with '||
True', avoiding a failure in case tarball does not come with an
installer 5. Remove installer if present.

An example of the produced dockerfiles from this new approach, can be
seen at [1].

[1] https://review.trustedfirmware.org/c/ci/dockerfiles/+/9109

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I1c5e67c636a6b36556c81be1407eccaa58395fe0