scp scmi-v2: Build scmi updated to use scmi-v2
build_package build_scmi_lib_scp_tools has been updated to work with
scmi-v2 as the build command has changed for this version.
The clock and sensor protocols have been excluded as they could not run
on hardware.

In addition, the branch for clone_tf_for_scp_tools is updated.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Co-authored-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: Ib22ab4571f2d8b9ae2f8965b51d8bb15cfb6011c