Juno: Upgrade SCP binaries to v2.8.0 for Juno
SCP v2.8.0 release depends on TF-A patch:
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/9557.
However, for that patch to be merged the Juno
SCP binaries need to be upgraded to v2.8.0.

This patch thus upgrades the SCP binaries to
a dev version of v2.8.0 (thus the directory
name css_scp_2.8.0-dev). Once SCP v2.8.0
is officially released this patch will be
updated.

Change-Id: I5c91ee6a33292a0853bc050f1bc7e10f9e3b72db
Signed-off-by: Zelalem <zelalem.aweke@arm.com>