fix(fvp): adjust lava-test-monitor timeout for fvp TFTF jobs
Curently, there is no timeout value for actions' lava-test-monitor,
thus LAVA defaults it to 30 seconds. This is an sensitive value
which defines the  maximum allowed time for the testing phase (test
action, through monitors), with the consequence that currently LAVA
only test for 30 seconds [1], a value too short for TFTF testing. This
commit, defines this value to 45 minutes, enough time for all cases
seen so far, i.e. [2]. Also, it explictitly defines other relevant
timeout values, placing all these on top of job definition.

To take [1] and [2] as job examples, [1] within 30 seconds, is only
able to execute 3 tests, while [2] 137.

[1] https://tf.validation.linaro.org/scheduler/job/209121
[2] https://tf.validation.linaro.org/scheduler/job/207989

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I475eb7f5aae28e50355a2a25881c21aa5c588e99