feat(coverity-scan):Notify if newly-added files not included for coverity scan
Added a test script to detect the newly added files between patches,
not analysed by the tf-coverity job. The script checks for only source
files("*.c") and logs the filenames not included for coverity analysis.
The test is included along with other existing static checks and
executed as part of 'tf-static-checks' under level 1 CI Job -
"tf-gerrit-tforg-l1".

Signed-off-by: Jayanth Dodderi Chidanand <jayanthdodderi.chidanand@arm.com>
Change-Id: I846d56010dae521f375515d993855a648635ef81