fix(common_def): use distclean in clean_build()
Just doing a "make clean" in the clean_build() function does not remove
all previous build artefacts, as the build/<platform> directory stays
around. This leads to build errors when files get added or removed
between builds because of changed build options.

Replace the "make clean" with "make distclean" to cover those cases,
also drop the not needed (and actually harmful) options to "make" on the
way.

Change-Id: Ie8fc2223606d2c74de4dbe3f7836a6d85206986d
Signed-off-by: Andre Przywara <andre.przywara@arm.com>