fix(tc): fix SCP binaries name
In those configurations where SCP is built, the binaries are
collected and stored into the archive for later use. During this
process, their names are changed and the "fw" at the end is removed.

Total Compute platforms still use the original name in
their run_config. This means that the binaries are not found in the
archive and the pre-built ones are used instead.

This patch fixes the name in the run_config for TC platforms such that
the fresh-built SCP binaries are used if part of the test.

Change-Id: I99b030fed539c46054bf9e1fe82c64be91ac558b
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>