gen_test_desc.py: Create .testprop files in parallel with .test
These files have Jenkins property file format, specifically with
keys TEST_CONFIG (original test config) and TEST_DESC
(expanded/normalized test config). Using them will allow better
interfacing with Jenkins job parameters. Currently, a .test file
is posted as "binaryfile" parameter where file name encodes
TEST_DESC, while the content - TEST_CONFIG. Doing it this way
is not entirely clear to someone not initimately familiar with
Jenkins details (e.g. it's not easy to suppose that when passing
a file parameter, Jenkins passes not just the content, but also
a verbatim file name). Besides, passing it this way then requires
extra clicks to the file content (and again, only content is visible,
not the filename, which makes it all confusing).

This change will allow to pass both values as normal string parameters,
easily and clearly visible in Jenkins UI.

For now, value of TEST_DESC matches the previous filename, to minimize
corresponding changes to the tf-a-builder job which parses it. It may
be cleaned up in the future (to be more human-readable).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ifa07e159701dd2462fdcab3c78d07161676ea6f5