feat(plat/fvp_r): add run and test configs to boot Yocto Linux on FVP-R
This patch adds the appropriate files to build TF-A for fvp_r
and then boot to the Yocto Linux command line. If the command
line is reached the test is considered successful.

This patch only adds support for ARM internal Jenkins CI, support for
OpenCI will be added in a future patch.

Files needed by Jenkins can be found here:
http://files.oss.arm.com/downloads/tf-a/fvp_r/

Change-Id: Id18f76b0233d67a8ae8967fedc5b93220a09002e
Signed-off-by: John Powell <john.powell@arm.com>
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>