fix: install Python dependencies
This change adds a new script - `install_python_deps_tf.sh` - which
creates and loads a clean Python virtual environment and within it
installs TF-A's Python dependencies.

The previous method for doing this (sourcing the environment from
within the pre-build hook) was not working as the environment was going
out of scope after returning from the function.

Change-Id: I1c2e404665853d33212254ad576a8095b5c25aa0
Signed-off-by: Chris Kay <chris.kay@arm.com>