refactor(juno): streamline SCP binary copying
For Juno we need to copy the latest SCP BL1 binary into the recovery
image, which has to match the SCP BL2 we put into the TF-A FIP image.
However there is no need to copy that scp_bl2.bin into the recovery
image, it's just in the original recovery image for convenience reasons
(to allow simpler rebuilds). So drop this explicit copy from when we
re-generate the recovery image.
Also drop the redundant get_scp_bl_bin() function, we can just use
fetch_file() for that.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Change-Id: I4b3df2b99d36215cd1e6e3eb1024f43e67896f44