feat(juno): update Linux kernel and U-Boot
At the moment we use a Linaro 5.3.0 kernel and some Linaro v2019.07
U-Boot version for the Juno Linux testing. Not only are those versions
outdated, but the old kernel does not exercise all features of recent
TF-A (for instance SOCID and TRNG).
Switch to mainline U-Boot (v2021.10) and mainline Linux (v5.15.5). We
then use a Busybox based initrd that drops into a user prompt.

Instructions to build these binaries from scratch using upstream source
repository are available in the readme.txt file [1].

To avoid picking up something from USB or the network, provide a changed
U-Boot environment, which drops all other boot targets except the NOR
flash ("afs" in U-Boot). For this to work, we need to fix the U-Boot
environment flash address in the images.txt files.

[1] http://files.oss.arm.com/downloads/tf-a/linux_boot/

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Change-Id: Idbce5c9c78d5fa48e34804ea5e877bfa162686d7