group: remove morello boot tests temporarily
Morello platform is undergoing a rather invasive change wherein
the boot flow is now being aligned to TBBR. This entails changes
in all software stacks across the platform which primarily
include TF-A and the SCP firmware and are currently in the
process of landing in their respective usptreams.

In order to get around the temporary CI failures arising out
of the synchronisation between TF-A and SCP upstreaming work,
it makes sense to disable the respective boot tests -
tf-l3-boot-tests-css and scp-boot-tests - for the Morello
platform until the new boot flow lands in the upstream branches.

These test configs shall be re-introduced again along with the
changes in their associated build/run configs to align with the
TBBR boot mechanism.

Signed-off-by: Anurag Koul <anurag.koul@arm.com>
Change-Id: I507e39c37a02dc866e724a08f797ea39c1d61ec5