ci(fvp-aarch32): increase temporarily the timeout of tftf expect script
Test configuration [1] taking more time to do the authentication of
images, and as a result of that expect script not able to find the
expected string in stipulated timeout i.e. 600secs
Increasing this timeout to 800secs helps to resolve this issue.
This is only interim fix, need to investigate further on why this
test configuration taking longer time to do the authentication.

[1]:tf-l3-boot-tests-misc/fvp-aarch32-tbb-mbedtls-ecdsa,fvp-aarch32-default:fvp-tftf.aarch32-fip.tftf-aemv8a.aarch32-debug

Change-Id: I01529edf92025e4fc1919cb12c9cc17587f6ec31
Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>