ci(expect):increase temporarily the timeout of tftf lava-expect script
Test configuration [1] taking more time to do the authentication of
images, and as a result of that expect script not able to find the
expected string in stipulated timeout i.e. 7 mins.
Increasing this timeout to 15mins helps to resolve this issue.
This is only interim fix, need to investigate further on why this
test configuration taking longer time to do the authentication.

[1]:tf-l3-boot-tests-misc/fvp-aarch32-tbb-mbedtls-ecdsa,fvp-aarch32-
default:fvp-tftf.aarch32-fip.tftf-aemv8a.aarch32-debug

Change-Id: I36a9c2a3354034923c0ec6a8971d2e6cc47946f8
Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>