group: remove n1sdp boot tests temporarily
N1SDP platform is undergoing a rather invasive change wherein
the boot flow is now being aligned to TBBR. This entails changes
in all software stacks across the platform which primarily
includes TF-A and the SCP firmware, and are currently in the
process of landing in their respective usptreams.

In order to get around the temporary CI failures arising out
of the synchronisation between TF-A and SCP upstreaming work,
it makes sense to disable the respective boot tests -
tf-l3-boot-tests-n1sdp and scp-boot-tests - for the N1SDP
platform until the new boot flow lands in the upstream branches.

These test configs shall be re-introduced again along with the
changes in their associated build/run configs to align with the
TBBR boot mechanism.

Signed-off-by: sahil <sahil@arm.com>
Change-Id: I35f66f31aaf105a956afce1425b8b1fc4795996c