fix(local-ci): fix group testing
It used to be possible to run an entire test group in the following
manner:

  $ test_groups=tf-l1-build-fvp run_local_ci.sh

This feature appears to be broken right now. Presumably it has been
broken since commit 0a9a3487090383bc5 ("spm: add support for new
hafnium/spm repo"), which changed the name of the test descriptor by
adding one digit.

For example,
  000%tf-l1-build-fvp%fvp-rdn1edgex2-tbb,nil,nil,nil,nil:nil.test
became
  0000%tf-l1-build-fvp%fvp-rdn1edgex2-tbb,nil,nil,nil,nil:nil.test

The trouble is that run_local_ci.sh was not updated in tandem at the
time and so it still assumed that test descriptors would be prefixed
by 3 digits. Down the line, this confused the test selection logic and
made the script fail.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I51db628cef39fb33f94b30cf93385f2837dac611