Upgrade to Mbed TLS 2.28.0.
Version 2.28.0 is the latest and greatest 2.x release of Mbed TLS
library.

Note that the Mbed TLS project published version 3.x some time
ago. However, as this is a major release with API breakages, upgrading
to this one might require some more involved changes in TF-A, which we
are not ready to do. We shall upgrade to mbed TLS 3.x after the v2.7
release of TF-A.

Also update the URL to mbedtls repository on Github -
https://github.com/ARMmbed/mbedtls redirects to
https://github.com/Mbed-TLS/mbedtls, so we might as well directly use
the latter.

Change-Id: I812724c6794d48a66e10b454fe5cd872f5fbb245
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>