fix(allwinner): build sun50i_a64 port with DEBUG=0
The Allwinner A64 port is notoriously short of SRAM space, and we run
into the limit regularly. Besides when addding new features, or
increasing the code size due to changes in generic code, this is also
sensitive to the toolchain used. A recent change for instance compiled
fine with Ubuntu's GCC 9.4.0, and with some GCC 11.2.0 build, but failed
with the CI's GCC version of "10.3-2021.07" from some Arm server.

To avoid this issue randomly blocking patches, let's produce RELEASE
builds for the A64 port only. This decreases the image size enough to
stay within the limit, and should give us some peace of mind for a while.

Also add a build command to enable the new Allwinner CPU idle DTB code,
to get this code covered by the CI.

Change-Id: I93eb36ed20dfc3d28b14ba6de950a78eec101137
Signed-off-by: Andre Przywara <andre.przywara@arm.com>