juno: Provide board revision explicitly for SCMI test
Currently we perform the SCMI test on any of the Juno board
available in the farm, r0, r1 or r2.

Because the majority of boards available in the farm are
revision 0 and the tests on revision 1 have been proven to
be unstable, we explicitly choose the revision 0 for our
SCMI tests.

This will ultimately allow a smooth operation of the CI
without affecting the robustness of the test.

Change-Id: I2dd620412a47cf0d5ac4ff80dc8e2aa28e88ae47
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>