ci(bl2_el3): update bl2 base address
Created a separate run configuration for testing bl2_at_el3
with PIE enabled.
BL31 images require more space with bl2_at_el3 with PIE enabled.
As a result, the BL2 base was shifted by 1 page (4096 bytes) so
there would be adequate space for BL31 in trusted SRAM.

This change is in-lined with TF-A [1].

[1]: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/15486

Change-Id: I4b04e7fbe9b7d5b70cbda3250c3f61bf5c9bacc2
Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>