fix(fvp): Revert "feat(fvp): improve ROMlib's jump table index file management"
This reverts commit 6b693e4fcf5b6e7202701613b42d4f3289b8b600.

This patch makes the following test configs fail:
* spm-l2-boot-tests/fvp-default,fvp-spm-tbb,fvp-default:fvp-spm+romlib
* spm-l2-boot-tests/fvp-default,fvp-spm-tbb-dualroot,fvp-default:fvp-spm+romlib
* tf-l2-boot-tests-misc/fvp-tbb-mbedtls-romlib:fvp-linux-dtb.1t-romlib-fip.uboot-aemv8a-debug

with the following error message:

  Assembler messages:
  Error: can't open /work/workspace/workspace/tf-worker_ws_10/trusted_firmware/build/fvp/debug/libwrapper/fdt_getprop.s for reading: No such file or directory

I suspect the patched index file is removed too soon, compared to when fvp_romlib_cleanup() was
called previously. This needs more investigation. Reverting the patch to fix the CI in the short
term. Will push a new version soon.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I31bc56edb980e626e2aede2c4dce7d90b76e2e89