fvp: add build test for measured boot with RSS
Introduce a new build test which overrides PLAT_RSS_NOT_SUPPORTED
build flag (which defaults to 1 on the Base AEM FVP) from the command
line. This will allow detecting any build regressions in future
development of the RSS communication driver.

Please note that this is purely an artifact to build the RSS driver
code as part of TF-A CI. RSS is not supported on FVP right now (or on
any other upstream platform, for that matter) so the resulting
firmware will not be functional.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I7616c3e080b6964bd5821ca3d078389a820d2621