feat(fvp): drop usage of AArch32 specific DTs
There should be no difference in a devicetree between an OS using AArch32
and one using AArch64.
In preparation for unifying the DT generation in the TF-A tree, drop the
usage of the AArch32 specific version already. This should not have a
real effect on a half-way recent OS, and allows to remove those AArch32
specific DT files from the TF-A tree completely.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Change-Id: Id126766bd2065f01bcb78b28e70a05b45cab48ae