tc2: capture RSS UART output
On TC2 FVP, RSS firmware prints messages on a third UART (in addition
to the "usual" UARTs used by AP and SCP firmwares). Add support for
monitoring this extra UART.

This means its output will be captured in a text file, which can be
later examined offline from the artefacts directory.

More importantly, this will make sure that some process consumes the
data produced on this UART, which if it wasn't, could make the FVP
stall (we've faced this issue before).

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I9c13b2670ddefa85b4277432c692a21d99563d36