upgrade to Mbed TLS 2.28.1
In anticipation of the next Trusted Firmware release update to the
newest 2.x Mbed TLS library.

Note that the Mbed TLS project publised version 3.x some time ago.
However, as this is a major release with API breakages, upgrading to
this one might require some more involved changes in TF-A, which we
are not ready to do. We shall upgrade to Mbed TLS 3.x after the v2.8
release of TF-A.

Also remove setting mbedtls_archive in arm-env.sh to tfa_downloads since
we now rely less and less on the Internal CI so having a local copy of
Mbed TLS on arm internal servers is no longer required.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ifce20bb716a2fa9ab089dde4bc2bf1cfeadf178d