refactor: factor Gerrit metadata generation into templates
Since the addition of the `expand_template` utility function to our Bash
utilities, we can expand Bash expressions in arbitrary files. In
particular, this allows us to generate YAML expressions within our LAVA
job templates, and begin reducing the amount of external script-based
templating we do.

This change begins with the Gerrit metadata section shared by both of
the FVP LAVA job templates, by removing the template-specific scripts
which generate the relevant YAML, and integrating it directly into the
templates.

Change-Id: I1e28791069e8194f967cb5b0df12a19db2a0f590
Signed-off-by: Chris Kay <chris.kay@arm.com>