refactor: use Bash arrays to generate deployment artefact
This change simplifies generation of the LAVA job YAML deployment
artefacts section.

Currently, this section is generated by `fvp_utils.sh` by modifying and
removing lines from the template file when certain binaries appear in
the model parameters, involving a lot of intricate find/replace.

Instead, this has been replaced with an associative array providing the
YAML key for each binary and the regular expression to search the model
parameters file for. If a binary appears in the model parameters file,
it is added to the list of artefacts for the current job. An expression
inside the job template file loops over each known artefact for the job
and generates the relevant YAML.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I873788e2ae486a549e4964984713377485314f4b