refactor: don't introduce LAVA macros unnecessarily
The existing mechanism for templating the LAVA job YAML is to use macros
of the format `{XYZ}`, similar to those accepted by LAVA for certain
data, and to do a find/replace on them.

This final change replaces this scheme with Bash variable expansion now
used for all other templated values.

Be aware that the model parameters file contains paths, which do still
need to be replaced with the LAVA job YAML syntax. As such, only places
that used the syntax and didn't need to no longer do.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I47d0b749e7d67185bbb9a23420f6d831f36dd0e2