script/run_package.sh: kill dangling models in cleanup
Internal CI suffers from instability due to dangling processes that are
missed during FVP run cleanup. Give models time to print user statistics
before terminating the process and killing their children. As the
infrastracture is in the process of being sunsetted, this need not be
sophisticated and can be removed once migration to OpenCI is complete.

Change-Id: I507bbe6bcdf85b15af40f58d002fac62a5e12bac
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>