ci(sme): add build and run configs for SME
TF-A Tests have SME tests but no config in the CI exercise them. This
patch adds an SME config.

Unlike the tests with SVE, SME is not tried with SPM, because SPMD
prohibits SME in Non-Secure world with it enabled.

The CI compiler does not support Armv8.7 so build for Armv8.6 and enable
the features required for 9.2 manually.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: If434448da87eda18d3d3995b6eedbe2dd201e98d