fix(expect): update crash test expectations
This change updates the crash test expectations to reflect the real
output of their tests, as they've become outdated. We haven't noticed
because these scripts were not attached to the payload UART of any
tests, which is the only situation in which the result of the Expect
script could determine the success of the test.

This change also ensures that tests using these scripts have them set as
the payload script, as opposed to the timeout scripts they typically
use.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: Ic9870a0aa6a191cdb6b02fcedeaea0490377704c